moved to the proper package and renamed the classes implementing CommandSender
