+++ bungee/src/main/java/com/wizardlybump17/wlib/bungee/command/BungeeCommandExecutor.java
@@ -1,6 +1,6 @@
 package com.wizardlybump17.wlib.bungee.command;
 
-import com.wizardlybump17.wlib.bungee.command.sender.CommandSender;
+import com.wizardlybump17.wlib.bungee.command.sender.BungeeCommandSender;
 import com.wizardlybump17.wlib.command.CommandManager;
 import com.wizardlybump17.wlib.command.exception.CommandException;
 import com.wizardlybump17.wlib.command.holder.CommandExecutor;
@@ -18,14 +18,14 @@ public class BungeeCommandExecutor extends Command implements CommandExecutor {
     }
 
     @Override
-    public void execute(com.wizardlybump17.wlib.command.CommandSender<?> sender, String commandName, String[] args) throws CommandException {
+    public void execute(com.wizardlybump17.wlib.command.sender.CommandSender<?> sender, String commandName, String[] args) throws CommandException {
         manager.execute(sender, commandName + " " + String.join(" ", args));
     }
 
     @Override
     public void execute(net.md_5.bungee.api.CommandSender sender, String[] args) {
         try {
-            execute(new CommandSender(sender), getName(), args);
+            execute(new BungeeCommandSender(sender), getName(), args);
         } catch (CommandException e) {
             manager.getHolder().getLogger().log(Level.SEVERE, "Error while executing a command", e);
         }
