+++ commands/src/main/java/com/wizardlybump17/wlib/command/args/reader/ArgsReader.java
@@ -1,6 +1,6 @@
 package com.wizardlybump17.wlib.command.args.reader;
 
-import com.wizardlybump17.wlib.command.CommandSender;
+import com.wizardlybump17.wlib.command.sender.CommandSender;
 import lombok.NonNull;
 import org.jetbrains.annotations.Nullable;
