+++ commands/src/main/java/com/wizardlybump17/wlib/command/completer/ArgumentCompleter.java
@@ -1,6 +1,6 @@
 package com.wizardlybump17.wlib.command.completer;
 
-import com.wizardlybump17.wlib.command.CommandSender;
+import com.wizardlybump17.wlib.command.sender.CommandSender;
 import lombok.NonNull;
 
 import java.util.List;
