+++ commands/src/main/java/com/wizardlybump17/wlib/command/executor/CommandExecutor.java
@@ -1,7 +1,7 @@
 package com.wizardlybump17.wlib.command.executor;
 
-import com.wizardlybump17.wlib.command.CommandSender;
 import com.wizardlybump17.wlib.command.exception.CommandException;
+import com.wizardlybump17.wlib.command.sender.CommandSender;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Map;
