+++ commands/src/main/java/com/wizardlybump17/wlib/command/holder/CommandExecutor.java
@@ -1,8 +1,8 @@
 package com.wizardlybump17.wlib.command.holder;
 
-import com.wizardlybump17.wlib.command.CommandSender;
 import com.wizardlybump17.wlib.command.args.reader.ArgsReaderException;
 import com.wizardlybump17.wlib.command.exception.CommandException;
+import com.wizardlybump17.wlib.command.sender.CommandSender;
 
 /**
  * Interface for intercepting commands when they are called
