+++ commands/src/main/java/com/wizardlybump17/wlib/command/sender/CommandSender.java
@@ -1,4 +1,4 @@
-package com.wizardlybump17.wlib.command;
+package com.wizardlybump17.wlib.command.sender;
 
 import org.jetbrains.annotations.NotNull;
