+++ core/src/main/java/com/wizardlybump17/wlib/command/reader/BlockDataArgsReader.java
@@ -1,8 +1,8 @@
 package com.wizardlybump17.wlib.command.reader;
 
-import com.wizardlybump17.wlib.command.CommandSender;
 import com.wizardlybump17.wlib.command.args.reader.ArgsReader;
 import com.wizardlybump17.wlib.command.args.reader.ArgsReaderException;
+import com.wizardlybump17.wlib.command.sender.CommandSender;
 import lombok.NonNull;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
